+++ PoliceApp/mobile/lib/core/network/dio_client.dart
@@ -1,4 +1,5 @@
 import 'package:dio/dio.dart';
+import 'package:flutter/foundation.dart';
 import '../constants/api_constants.dart';
 import '../storage/secure_storage_service.dart';
 import 'auth_interceptor.dart';
@@ -28,14 +29,19 @@ class DioClient {
       ),
     );
 
-    dio.interceptors.addAll([
-      AuthInterceptor(dio: dio, storage: storage),
-      LogInterceptor(
-        requestBody: true,
-        responseBody: true,
-        logPrint: (obj) => print('[HTTP] $obj'),
-      ),
-    ]);
+    dio.interceptors.add(AuthInterceptor(dio: dio, storage: storage));
+
+    // SECURITY: only log request/response bodies in debug builds.
+    // Release logs would leak passwords + access/refresh tokens to device logs.
+    if (kDebugMode) {
+      dio.interceptors.add(
+        LogInterceptor(
+          requestBody: true,
+          responseBody: true,
+          logPrint: (obj) => debugPrint('[HTTP] $obj'),
+        ),
+      );
+    }
 
     return dio;
   }